Moved out color scheme & syntax tests from view
