+++ lib/editor_test.go
@@ -10,6 +10,8 @@ import (
 
 	"github.com/limetext/lime-backend/lib/keys"
 	"github.com/limetext/lime-backend/lib/packages"
+	"github.com/limetext/lime-backend/lib/render"
+	"github.com/quarnster/parser"
 )
 
 func TestGetEditor(t *testing.T) {
@@ -197,8 +199,24 @@ func TestHandleInput(t *testing.T) {
 	}
 }
 
-func TestAddColorScheme(t *testing.T) {
+type dummyColorSc struct {
+	name string
+}
 
+func (d *dummyColorSc) Name() string {
+	return d.name
+}
+
+func (d *dummyColorSc) Spice(*render.ViewRegions) render.Flavour { return render.Flavour{} }
+
+func TestAddColorScheme(t *testing.T) {
+	cs := new(dummyColorSc)
+	ed := GetEditor()
+
+	ed.AddColorScheme("test/path", cs)
+	if ret := ed.colorchemes["test/path"]; ret != cs {
+		t.Errorf("Expected 'test/path' color scheme %v, but got %v", cs, ret)
+	}
 }
 
 func TestGetColorScheme(t *testing.T) {
@@ -209,8 +227,34 @@ func TestColorSchemes(t *testing.T) {
 
 }
 
-func TestAddSyntax(t *testing.T) {
+type dummySyntax struct {
+	name      string
+	filetypes []string
+	data      string
+}
 
+func (d *dummySyntax) Name() string {
+	return d.name
+}
+
+func (d *dummySyntax) FileTypes() []string {
+	return d.filetypes
+}
+
+func (d *dummySyntax) SetData(s string) {
+	d.data = s
+}
+
+func (d *dummySyntax) Parse() (*parser.Node, error) { return nil, nil }
+
+func TestAddSyntax(t *testing.T) {
+	syn := new(dummySyntax)
+	ed := GetEditor()
+
+	ed.AddSyntax("test/path", syn)
+	if ret := ed.syntaxes["test/path"]; ret != syn {
+		t.Errorf("Expected 'test/path' syntax %v, but got %v", syn, ret)
+	}
 }
 
 func TestGetSyntax(t *testing.T) {
